+++ src/clojure/pantomime/mime.clj
@@ -1,17 +1,11 @@
 (ns pantomime.mime
+  (:use pantomime.internal)
   (:import [java.io File InputStream]
            [java.net URL]
            [org.apache.tika Tika]
            [org.apache.tika.mime MediaType MimeType]))
 
 (def ^Tika detector (Tika.))
-
-;; clojure.java.io has these as private, so we had to copy them. MK.
-(def ^{:doc "Type object for a Java primitive byte array."}
-  byte-array-type (class (make-array Byte/TYPE 0)))
-
-(def ^{:doc "Type object for a Java primitive char array."}
-  char-array-type (class (make-array Character/TYPE 0)))
 
 ;;
 ;; API